ADDED: CrearCoordenadasMapa.cs  DELETED: Mapa*.txt
Añadida funcionalidad para controlar los poderes de los jugadores y la ubicación de estos en los distintos mundos.
Se ha cambiado la implementación de la carga de coordenadas por mapa.
Ya no son necesarios los txt, esto evita tener que pasar las coordenadas manualmente de todos los mapas a los txt. Se automatiza el proceso.
Ahora es necesario INCLUIR TODOS LOS SPAWN en su gameobject de la jerarquía de Unity para que CrearCoordenadas tenga control sobre ellos.
CrearCoordenadas se ocupa de recopilar las coordenadas de todos los spawns, informar a los componentes y destruir las referencias públicas cuando ya se ha tomado control de toda la información.
Componente PuntoInicialMapa ahora se destruye tras informar a GameManager y que este coloque a los jugadores.
Retocado algunos scripts para mejorar gestión de información.
Se han modificado algunos métodos que se usaban por duplicado.
Ahora GameManager solo asigna la victoria al primer jugador.
Reasignado todos los poderes en BuscaHabilidad.
Muro vuelve a ser funcional y se instancia en su debido mundo/posición.
Neblina vuelve a estar implementada desde 1 solo prefab.
Neblina ahora se instancia y se destruye tras el tiempo indicado en su debido mundo/posición con su debido tamaño.
Se ha sustituido la escena "Mapa1" por "test".
Se ha mejorado la visualización de la carga de los mapas, ahora los jugadores permanecerán ocultos y desactivados, así como sus controles hasta que sean ubicados en sus posiciones iniciales de cada mundo.
Se ha reducido el tiempo de la carga.

+++ Assets/Scripts/Enemigos/GoombaStun.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class GoombaStun : MonoBehaviour {
 
-    public int segundos;
+    public float segundos;
     
 
     /// <summary>
